app/lambda/ingest/requirements.txt
ccxt==4.3.6
pandas==2.2.2
numpy==1.26.4
google-cloud-pubsub==2.21.1
boto3==1.34.0
google-auth==2.29.0